space/ai-training/requirements.txt
torch>=1.9.0
transformers>=4.30.0
datasets>=2.12.0
sentencepiece>=0.1.99
jieba>=0.42.1
nltk>=3.8.1
mlflow>=1.20.0
pyyaml>=5.4.1
tqdm>=4.62.0
numpy>=1.21.0
psutil>=5.8.0
tensorboard>=2.6.0
pandas>=1.5.0
scikit-learn>=1.0.0
pillow>=8.3.1
tabulate>=0.8.9